svg/extensibility/interfaces/foreignObject-graphics.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><metadata><link xmlns="http://www.w3.org/1999/xhtml" href="https://www.w3.org/TR/SVG2/embedded.html#InterfaceSVGForeignObjectElement" rel="help"/><link xmlns="http://www.w3.org/1999/xhtml" href="https://svgwg.org/svg2-draft/changes.html#extend" rel="help"/><meta xmlns="http://www.w3.org/1999/xhtml" content="foreignObject is a graphics element" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script><![CDATA[
  test(function() {
    var target = document.getElementById('target');
    assert_equals(target.__proto__.__proto__, SVGGraphicsElement.prototype);
  });
  ]]></h:script></svg>
